pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
streamlit = "^1.15.2"
audio-recorder-streamlit = "^0.0.4"
"whisper.ai" = "^1.0.0"
pydub = "^0.25.1"
streamlit-audiorecorder = "^0.0.2"
replicate = "^0.4.0"
pyaudio-wheels = "^0.0.1"
websockets = "^10.4"
keyboard = "^0.13.5"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"